xai-engine-service/requirements.txt
# === CORE FRAMEWORK ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# === DATABASE & ORM ===
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# === CELERY & REDIS ===
celery==5.3.4
redis==5.0.1

# === XAI LIBRARIES ===
shap==0.43.0
lime==0.2.0.1

# === ML & DATA SCIENCE ===
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
pyarrow==14.0.1

# === OPENAI API ===
openai==1.51.2
tiktoken==0.8.0
httpx==0.27.2

# === UTILITIES ===
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0
python-jose[cryptography]==3.3.0

# === STORAGE CLIENTS ===
minio==7.2.0
azure-storage-blob==12.19.0

# === MONITORING & LOGGING ===
loguru==0.7.2
